Separate Entity/ folder (and BC)
